fix: merge missing output routes tests for pkg-file-count
Splitting the table test into missing_output_routes_test.go satisfied
backend-size but exceeded the validation package file cap (16/15).
Keep helpers and cases in validation_test.go with a short test function.
